REFS #66 * Move the service to an independent file.
