+++ minos/networks/broker/services.py
@@ -1,0 +1,36 @@
+"""
+Copyright (C) 2021 Clariteia SL
+
+This file is part of minos framework.
+
+Minos framework can not be copied and/or distributed without the express permission of Clariteia SL.
+"""
+from __future__ import (
+    annotations,
+)
+
+from aiomisc.service.periodic import (
+    PeriodicService,
+)
+from minos.common import (
+    MinosConfig,
+)
+
+from .dispatcher import (
+    MinosQueueDispatcher,
+)
+
+
+class MinosQueueDispatcherService(PeriodicService):
+    """Minos QueueDispatcherService class."""
+
+    def __init__(self, config: MinosConfig = None, **kwargs):
+        super().__init__(**kwargs)
+        self.dispatcher = MinosQueueDispatcher.from_config(config=config)
+
+    async def callback(self) -> None:
+        """Method to be called periodically by the internal ``aiomisc`` logic.
+
+        :return:This method does not return anything.
+        """
+        await self.dispatcher.dispatch()
